Rename LeagueItemAccountIdList.java to LeagueItemMap
